Setup user accounts, mongoose, splash, session
Setup user account creation and login. Configured sessions and mongoDB. Changed main page to be a login splash page if you aren't logged in and it will show the original home page once you are logged in. Got a nav bar setup with a logout option. Setup routes and controllers for registration and login. Created a form validation and catchAsync middleware. Added another webpack for the splash page that contains the form validation.