Changes based on Jake's feedback
Added mongo host and port to .env and added them to the line for the mongodb connection.

Changed the require paths for the user model and routes to exclude the server folder since it was unnecessary.

Moved the navbar to a separate header file and imported that into the main.js and used it in place of where it had been previously. Also created a components folder in case we make more in the future.

Uninstalled ejs since we are using ejs-mate due to ejs not knowing how to handle layouts natively.