Dockerized the application
Docker lets you specify an environment configuration and it'll handle installing all of your dependent programs and setting it up with some slick networking stuff. It'll also automatically restart stuff if your server goes down!

I took what I've done on other projects and modified it a little. There are a few things of note here. The `Dockerfile` tells docker-compose how to build and run the application, and you run it by running `./run.sh dev` or `./run.sh prod` (on linux, or WSL, with docker and docker compose installed). It'll first check that you have a `.env` file, otherwise it'll copy the sample for you so you can fill it out and try running again. The `run.sh` file is necessary because the configuration is broken up into some additional `.dev.yml` and `.prod.yml` docker compose files. This lets us specify different things (like different restart policies, so in production it'll always restart your server, but it lets you stop it in development) and in development it can use a bind mount to get live file changes (and run them with nodemon) without having to restart the docker container, versus production which will bake in the files. I didn't try out the production setup but it should more or less work as it's basically what I do everywhere else :P.

I added the docker defaults to the `.env` file, so you don't need to figure out basic docker networking and can just specify a session secret (so far) and not have to worry about configuring the connection string. But if you're using a non-docker MongoDB you can just change the host to `localhost` and the port to whatever you need.

I added `nodemon` to the dev dependencies, so we can run `nodemon` commands from `npm` scripts or with `npx` without having to install it globally.