Added script content hashing
When you load a page you want the browser to cache things. I haven't added settings (http headers) to cache things yet (we will do that someday when we switch from express serving assets with `express.static()` to using nginx as a reverse proxy) but this sets us up for proper cache busting.

Cache busting is the term for how you tell the browser scripts need to be loaded again. You watch scripts to be cached all the time so the browser doesn't have to ask your server for them every single time a user loads the page, so you want a really long cache expiration time and some way to bust it. This is the strategy I've been using on my other projects. I installed a new npm package "webpack-manifest-plugin" which will output a `manifest.json` file that maps the original script file names to their new names with a hash of the contents in the file name. That way whenever that hash in the file name changes then the browser will see the new script src in the html (of pages that load after the server updates) and say "oh I don't have main.a7cae40596be273f238b.js cached, better ask the server for it!"

The manifest file looks like `{ "main.js": "main.a7cae40596be273f238b.js" }`, with additional properties for any other scripts in there. The content hash is generated automatically by webpack because the webpack config now includes `[contenthash]` in the output file names. The final key to this is that the server must always send the newest content hashed file name! This is where that manifest comes in handy. There's a new `getManifest` utility function used by all of the views that will load the manifest and include it in the page locals. Then in the EJS templates we can get the script file names we want by accessing the desired entry point script from the manifest, so on `home.ejs` the script changed from `src="/main.js"` to `src="<%= manifest['main.js'] %>"` so it'll dynamically insert whatever the newest script name is.

The one final gotcha is the way the manifest is requested. You can use `require()` to load a json file, like the manifest, but `require` will cache the result. So in that case you'd need to restart the server before you would see any of your changes or it'd keep trying to serve you old scripts. I fixed this by only using `require()` for production so it can use the cached manifest, and to manually load and parse the file with `fs` every time the manifest is needed when in development. `getManifest` is an async function so now all the routes that needed it are wrapped in `catchAsync`.

Additionally I fixed the `nodemon` command so it will now ONLY restart the server if a script in `src/server` changes.

I also fixed some react warnings. It didn't like the usage of `class` in the Header component, it wants you to use `className`, an honest mistake!