To-Do Lists and Prisma Refactor
Switched from using Mongo to Prisma due to needing the relational db more than Mongo.

Added functionality with IGDB and they advise to run a local copy of their db and use webhooks to get updates to it. Haven't setup the update part but that will come later. For now you will run the file for creating the local copy of the igdb database if changes are needed.

Added todo lists that are assigned to the user that created the list. This will also only show to that user if they are signed in. You can create lists and then add games to the lists by searching for a game and selecting the desired game in the drop down. Styling needs to be fine tuned still. Can click on the x to delete the item/list if needed.