Fixed conditional for rendering the search results and updated schema
Fixed the conditional for rendering the search results to reflect an earlier change to how I set the searchResult state if the searchResult state is empty. Also realized I didn't have the list only allow for a game to be in there once so I added unique to the game table.