Added Random Button
Added a random button to show on each game list. When clicked, if there is more than 1 game in the list it will display a toast notification with a random game from that list. If there is only 1 game it will pop up and prompt to add another game first. If there are no games, it will say to add games to the list first.