Fixed deleting if there is a list on a game page
Fixed the issues that were present if you tried to delete anything if there was a list on a game page within the item you were trying to delete. Prisma 3 has a onDelete cascade that has been implemented now so I changed all my deletion to accommodate for that.

Changed another console.log to console.error that I missed that last time around.

Fixed adding a todo to a todo list since it wasn't working at all before do to me having incorrect properties I was looking for when I was trying to create the toDoItem.