Added play status
Added a play status to each game. There is a dropdown on the Game Page that you pick what the game's status is. By default it is planning to play.

Made the gamePlayStatus it's own component so it won't clutter the gamePage component file.

Also added next to the props of the controller for game.js since I realized it was missing to it would error out if those functions ever got an actual error.