Fixes for the gameStatus PR
Implemented an array for statuses that I filter and map over to make the dropdown that doesn't have the current status in it. Made a couple variables, filteredStatus and currentStatus that filter the statuses array to just being whatever the current play status is for the game.

Added styling to my scss file to change the look of the buttons to not be buttons and blend in better with the dropdown list.