Fixed changes from comments
Changed the filter to find for the currentStatus. Removed the key from the GamePlayStatus call on gamePage.js and added a key to the li's on the gamePlayStatus.js file.

Changed the classes on the buttons on the dropdown to be like how they have the Menu Items on the bootstrap docs. Removed the scss file for this since it was no longer needed afterwards.

Also realized that I wasn't passing in the parentGameId properly when I was creating or deleting lists so it was returning all lists since the parentGameId was undefined. Added a ternary operator to the query so if req.query.parentGameId is undefined then it will use req.body.parentGameId instead.