Added new tables to db and changed game search to FTS
Changed the db to be many to many relationships and added fields for more info to show on the game page. Also got alt names and use that with the full text search preview feature of prisma for more accurate search results. Removed the index from the prisma schema since one was manually created on the db since the FTS feature needs indexes to be managed manually.