Changes for PR convos
In igdbWebhook I made 2 functions, one for fetching and one for creating and replaced all but the creating games with those functions. The creating games is done differently so I left that one as is. Also removed all the count portions since they weren't needed anymore. Changed how the backend was sending the search results so I was able to revert the change made to the front end for that part. Also changed how the items such as player perspective and genres were being set in the igdbGame creation portion to be slightly shorter.