Changed input field handling and reworded things
Changed the New Game text in the game lists to say Add Game. Changed the To Do Lists input to say Add To-Do rather then New To Do for consistency.

The input boxes now clear themselves upon submitting the forms.

The game searching will now say Searching... while it is searching for games and will say No Results if it found nothing.

Renamed the site to be titled Backlog Ninja rather than Game Tracker.

Also tried to fiddle with the indexing for game names to improve searching.