Changes for PR
Changed the searchResult in controller back to const rather than let. This was from how I tried to handle the no results in the testing and was left over from an attempt at that.

Used React's ref for clearing out the search input after submitting rather than using the document.getElementById.

Removed the isSearching state change in handleChange and change it in the useEffect instead now.

Renamed gameForm to be searchFormInput so it is easier to understand what it is for.