Fix for 500 error when searching
If you would have tabs, newline, and trailing, leading or consecutive white spaces the server would throw a 500 error because it was trying to search as that character listed earlier.

I changed the way I was mutating the search string to use underscores for all the white spaces, tabs and new lines that may be in the search string since the search will act the same if they are underscores instead.