Improved search results, manual list sorting, visual updates, updating backend webhooks
Added manual list sorting using the react-beautiful-dnd package. To accommodate for this I had to update how I handle deleting games to update the list position when a game is deleted so there isn't a potential break in the numbering.

Set the searching to order the results by the relevance and then filters out games that are already in the list. It then takes the first 10 items of that filtered list rather than getting the first 10 from the database since there could be less than 10 results based on filtering.

Removed the drop shadow on the game search since I added a background on the games so they were solid when moving and that background cut into the shadow and looked weird.

Changed the account dropdown to have an icon rather than say account.

Updated the webhook to be able to get a new oauth token and  replace it in the env and also fixed the main getGames function to be able to get the new games between the last update and whenever you call that function.