Fix game search errors and hiding carousel errors
Updated node & npm on my computer since it was rebuilding the package lock on the hosted version since I was on an older version.

If you had special characters such as a colon in the game search it would error out so I changed the search input regex to replace them with an underscore and also trimmed the start and end of underscores that may be there after the first replacement.

Added a check for the hl2b times so it won't show the arrows if there are no game times at all or before the game time carousel is rendered.

Reduced padding on all lists and changed the game to do lists style to match the game lists.

There were also a couple things I was pretty sure I changed for webstorm warnings that seemed to change back for some reason so they were resolved again.