Making To-do's editable and reorder To-do's
Added the ability to edit to-do's using contenteditable on the spans for the to-do's. Removed the outline that it gives it since it was just a finely dotted line that didn't match the style of the rest of the page.

Added the ability to reorder the to-do's in the same vein as the game lists. Renamed some of the functions so it was easier to tell what they are for now that there are 2 different list types that can be reordered.

Changed styling on the input fields to remove the default focused color.

Changed the draggables to only be draggable from the dots on the left hand side of the element.

Tried to fix the game searching index issue as well but opted to just remove the indexing I had done since it appears to be a known working bug that it ignores indexes for FTS and they are currently working on resolving the issue. Once they have it fixed I may have to manually make the index on the DB if they don't add the ability for Prisma to make the index.