Changed home in navBar to use react-router
The home button in the navBar had been using the href that was part of react-bootstraps NavBar.Brand so it was reloading the whole page. Switched the the NavBar.Brand to be within a react-router link that points to '/' and then removed the href in NavBar.Brand. Also removed the text-decoration on the home link since it underlined it since it is using Link now.