Fixed game pages and updates packages
Going to a game page would break the site. The hl2b api changed how it works so the requests 404 with that. I need to find a better solution for it but in the mean time I commented the lines out so the game pages will at least work.

I also updated almost all the packages since they were starting to get pretty out of date. The only one I didn't update was redis because I wasn't able to get it to connect and load the page after updating it to the latest version.