Added visual change for completed games & new sort
If a game's status is set to completed it now has a strike-through on the text and is set to 50% opacity.

I also added a new sort that when selected will sort all games not set to completed and sort them at the top of the list by a-z and then sort all the completed games a-z and put them at the end of the list.

Updated the random game button to now not include games that have completed and dropped for the status.